ui/requirements.txt
streamlit 
typing_extensions==4.9.0
psycopg2-binary
langchain-groq
langchain-cohere
langchain
python-dotenv

